+++ pinata_client/components/game/ThrowBallPlay.tsx
@@ -1,9 +1,11 @@
 import React, { useEffect, useRef } from "react";
-import { Animated, PanResponder, View, StyleSheet, Image } from "react-native";
+import { Animated, PanResponder, View, StyleSheet } from "react-native";
 import LottieView from "lottie-react-native";
 import { useGame } from "@/hooks/useGame";
 import GameResultPopup from "./GameResultPopup";
 import { useAuthStore } from "@/stores/authStore";
+import Target from "./Target";
+import Ball from "./Ball";
 
 export default function ThrowBallPlay() {
   const ballY = useRef(new Animated.Value(0)).current;
@@ -11,28 +13,6 @@ export default function ThrowBallPlay() {
   const boxX = useRef(new Animated.Value(0)).current;
   const { throwBall } = useGame(lottieRef, ballY, boxX);
   const { user } = useAuthStore();
-
-  useEffect(() => {
-    const animateBox = () => {
-      const distance = Math.random() * 150 + 50; // 이동 범위 (50 ~ 200px)
-      const duration = Math.random() * 1000 + 1000; // 속도 (1 ~ 2초)
-
-      Animated.sequence([
-        Animated.timing(boxX, {
-          toValue: distance,
-          duration,
-          useNativeDriver: false,
-        }),
-        Animated.timing(boxX, {
-          toValue: -distance,
-          duration,
-          useNativeDriver: false,
-        }),
-      ]).start(() => animateBox()); // 반복
-    };
-
-    animateBox();
-  }, []);
 
   const handleReset = () => {
     ballY.setValue(0);
@@ -50,28 +30,8 @@ export default function ThrowBallPlay() {
 
   return (
     <View style={styles.playContainer}>
-      <Animated.View
-        style={[styles.boxWrapper, { transform: [{ translateX: boxX }] }]}
-      >
-        <LottieView
-          ref={lottieRef}
-          source={require("../../assets/animations/gift_lottie.json")}
-          autoPlay={false}
-          loop={false}
-          style={styles.lottie}
-        />
-      </Animated.View>
-      <Animated.View
-        {...panResponder.panHandlers}
-        style={[styles.ball, { transform: [{ translateY: ballY }] }]}
-      >
-        <Image
-          source={require("../../assets/images/ball.png")}
-          style={{ width: 30, height: 30 }}
-          resizeMode="contain"
-        />
-      </Animated.View>
-
+      <Target boxX={boxX} lottieRef={lottieRef} />
+      <Ball ballY={ballY} panHandlers={panResponder.panHandlers} />
       <GameResultPopup onReset={handleReset} />
     </View>
   );
@@ -82,19 +42,5 @@ const styles = StyleSheet.create({
     flex: 1,
     justifyContent: "flex-start",
     alignItems: "center",
-  },
-  boxWrapper: {
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  lottie: {
-    width: 320,
-    height: 320,
-  },
-  ball: {
-    position: "absolute",
-    bottom: 60,
-    left: "50%",
-    marginLeft: -15,
   },
 });
